Acomodo de tests

+++ src/main/java/dao/UsuarioDAO.java
@@ -2,7 +2,7 @@ package dao;
 
 import model.Usuario;
 
-public class UsuarioDAO extends GenericDAOImpl<Usuario, Integer> {
+public class UsuarioDAO extends GenericDAOImpl<Usuario, Long> {
     public UsuarioDAO() {
         super();
     }
